requirements.txt
fastapi
tortoise-orm
aerich
black
mypy
uvicorn
requests
pytest
pytest-asyncio
